+++ finite-volume/source/cuda/interpolation_v2.h
@@ -30,7 +30,7 @@
 #define READ(i)	__ldg(&read[i])
 
 template<int log_dim, int block_type>
-__global__ void interpolation_v2_kernel(level_type level_f, int id_f, double prescale_f, level_type level_c, int id_c, communicator_type interpolation){
+__global__ void interpolation_v2_kernel(level_type level_f, int id_f, REAL prescale_f, level_type level_c, int id_c, communicator_type interpolation){
   // one CUDA thread block operates on one HPGMG tile/block
   blockCopy_type block = interpolation.blocks[block_type][blockIdx.z];
 
@@ -55,8 +55,8 @@ __global__ void interpolation_v2_kernel(level_type level_f, int id_f, double pre
   int write_jStride = block.write.jStride;
   int write_kStride = block.write.kStride;
 
-  double * __restrict__  read = block.read.ptr;
-  double * __restrict__ write = block.write.ptr;
+  REAL * __restrict__  read = block.read.ptr;
+  REAL * __restrict__ write = block.write.ptr;
   if(block.read.box >=0){
      read_jStride = level_c.my_boxes[block.read.box ].jStride;
      read_kStride = level_c.my_boxes[block.read.box ].kStride;
@@ -68,11 +68,11 @@ __global__ void interpolation_v2_kernel(level_type level_f, int id_f, double pre
     write = level_f.my_boxes[block.write.box].vectors[id_f] + level_f.my_boxes[block.write.box].ghosts*(1+write_jStride+write_kStride);
   }
 
-  double c1 = 1.0/8.0;
+  REAL c1 = 1.0/8.0;
   for(int k=0;k<write_dim_k;k+=2){
-    double c1i=c1;if(i&0x1){c1i=-c1;}
-    double c1j=c1;//if(j&0x1){c1j=-c1;}
-    double c1k=c1;//if(k&0x1){c1k=-c1;}
+    REAL c1i=c1;if(i&0x1){c1i=-c1;}
+    REAL c1j=c1;//if(j&0x1){c1j=-c1;}
+    REAL c1k=c1;//if(k&0x1){c1k=-c1;}
     int write_ijk = ((i   )+write_i) + (((j   )+write_j)*write_jStride) + (((k   )+write_k)*write_kStride);
     int  read_ijk = ((i>>1)+ read_i) + (((j>>1)+ read_j)* read_jStride) + (((k>>1)+ read_k)* read_kStride);
     //
@@ -80,33 +80,33 @@ __global__ void interpolation_v2_kernel(level_type level_f, int id_f, double pre
     // |---+---|---+---|---+---|
     // |   |   |???|   |   |   | fine grid
     //
-    double r11 = READ(read_ijk-1-read_jStride-read_kStride);
-    double r12 = READ(read_ijk-read_jStride-read_kStride  );
-    double r13 = READ(read_ijk+1-read_jStride-read_kStride);
-    double r21 = READ(read_ijk-1             -read_kStride);
-    double r22 = READ(read_ijk             -read_kStride  );
-    double r23 = READ(read_ijk+1             -read_kStride);
-    double r31 = READ(read_ijk-1+read_jStride-read_kStride);
-    double r32 = READ(read_ijk+read_jStride-read_kStride  );
-    double r33 = READ(read_ijk+1+read_jStride-read_kStride);
-    double r41 = READ(read_ijk-1-read_jStride             );
-    double r42 = READ(read_ijk-read_jStride               );
-    double r43 = READ(read_ijk+1-read_jStride             );
-    double r51 = READ(read_ijk-1                          );
-    double r52 = READ(read_ijk                            );
-    double r53 = READ(read_ijk+1                          );
-    double r61 = READ(read_ijk-1+read_jStride             );
-    double r62 = READ(read_ijk+read_jStride               );
-    double r63 = READ(read_ijk+1+read_jStride             );
-    double r71 = READ(read_ijk-1-read_jStride+read_kStride);
-    double r72 = READ(read_ijk-read_jStride+read_kStride  );
-    double r73 = READ(read_ijk+1-read_jStride+read_kStride);
-    double r81 = READ(read_ijk-1             +read_kStride);
-    double r82 = READ(read_ijk             +read_kStride  );
-    double r83 = READ(read_ijk+1             +read_kStride);
-    double r91 = READ(read_ijk-1+read_jStride+read_kStride);
-    double r92 = READ(read_ijk+read_jStride+read_kStride  );
-    double r93 = READ(read_ijk+1+read_jStride+read_kStride);
+    REAL r11 = READ(read_ijk-1-read_jStride-read_kStride);
+    REAL r12 = READ(read_ijk-read_jStride-read_kStride  );
+    REAL r13 = READ(read_ijk+1-read_jStride-read_kStride);
+    REAL r21 = READ(read_ijk-1             -read_kStride);
+    REAL r22 = READ(read_ijk             -read_kStride  );
+    REAL r23 = READ(read_ijk+1             -read_kStride);
+    REAL r31 = READ(read_ijk-1+read_jStride-read_kStride);
+    REAL r32 = READ(read_ijk+read_jStride-read_kStride  );
+    REAL r33 = READ(read_ijk+1+read_jStride-read_kStride);
+    REAL r41 = READ(read_ijk-1-read_jStride             );
+    REAL r42 = READ(read_ijk-read_jStride               );
+    REAL r43 = READ(read_ijk+1-read_jStride             );
+    REAL r51 = READ(read_ijk-1                          );
+    REAL r52 = READ(read_ijk                            );
+    REAL r53 = READ(read_ijk+1                          );
+    REAL r61 = READ(read_ijk-1+read_jStride             );
+    REAL r62 = READ(read_ijk+read_jStride               );
+    REAL r63 = READ(read_ijk+1+read_jStride             );
+    REAL r71 = READ(read_ijk-1-read_jStride+read_kStride);
+    REAL r72 = READ(read_ijk-read_jStride+read_kStride  );
+    REAL r73 = READ(read_ijk+1-read_jStride+read_kStride);
+    REAL r81 = READ(read_ijk-1             +read_kStride);
+    REAL r82 = READ(read_ijk             +read_kStride  );
+    REAL r83 = READ(read_ijk+1             +read_kStride);
+    REAL r91 = READ(read_ijk-1+read_jStride+read_kStride);
+    REAL r92 = READ(read_ijk+read_jStride+read_kStride  );
+    REAL r93 = READ(read_ijk+1+read_jStride+read_kStride);
  
     // i  j  k
     write[write_ijk] = prescale_f*write[write_ijk] +
@@ -165,13 +165,13 @@ __global__ void interpolation_v2_kernel(level_type level_f, int id_f, double pre
   interpolation_v2_kernel<log_dim,block_type><<<grid,block>>>(level_f,id_f,prescale_f,level_c,id_c,interpolation);
 
 extern "C"
-void cuda_interpolation_v2(level_type level_f, int id_f, double prescale_f, level_type level_c, int id_c, communicator_type interpolation, int block_type)
+void cuda_interpolation_v2(level_type level_f, int id_f, REAL prescale_f, level_type level_c, int id_c, communicator_type interpolation, int block_type)
 {
   int num_blocks = interpolation.num_blocks[block_type]; if(num_blocks<=0) return;
   dim3 block = dim3(min(level_f.box_dim,BLOCKCOPY_TILE_I), BLOCKCOPY_TILE_J, 1);
   dim3 grid = dim3((level_f.box_dim+block.x-1)/block.x,1,num_blocks);
 
-  int log_dim = (int)log2((double)level_f.dim.i);
+  int log_dim = (int)log2((REAL)level_f.dim.i);
   switch(block_type){
     case 0: KERNEL_LEVEL(log_dim,0); CUDA_ERROR break;
     case 1: KERNEL_LEVEL(log_dim,1); CUDA_ERROR break;
