+++ finite-volume/source/cuda/stencils/chebyshev.flux.fv4.h
@@ -29,15 +29,15 @@
 //------------------------------------------------------------------------------------------------------------------------------
 template<int LOG_DIM_I, int BLOCK_I, int BLOCK_J, int BLOCK_K>
 //__launch_bounds__((BLOCKCOPY_TILE_I * BLOCKCOPY_TILE_J), RESIDUAL_MIN_BLOCKS_PER_SM)
-__global__ void chebyshev_kernel(level_type level, int x_id, int rhs_id, double a, double b, int s, double *chebyshev_c1, double *chebyshev_c2){
-      extern __shared__ double l_flux[];
+__global__ void chebyshev_kernel(level_type level, int x_id, int rhs_id, REAL a, REAL b, int s, REAL *chebyshev_c1, REAL *chebyshev_c2){
+      extern __shared__ REAL l_flux[];
       int block = blockIdx.x;
 
-      double * __restrict__ flux_i = l_flux;
-      double * __restrict__ flux_j = l_flux + (BLOCK_I+1)*(BLOCKCOPY_TILE_J+1)*(1);
-      double * __restrict__ flux_k = l_flux + (BLOCK_I+1)*(BLOCKCOPY_TILE_J+1)*(2);
+      REAL * __restrict__ flux_i = l_flux;
+      REAL * __restrict__ flux_j = l_flux + (BLOCK_I+1)*(BLOCKCOPY_TILE_J+1)*(1);
+      REAL * __restrict__ flux_k = l_flux + (BLOCK_I+1)*(BLOCKCOPY_TILE_J+1)*(2);
 
-      double bh2invSTENCIL_TWELFTH = b*STENCIL_TWELFTH/(level.h*level.h);
+      REAL bh2invSTENCIL_TWELFTH = b*STENCIL_TWELFTH/(level.h*level.h);
 
 
       const int box  = level.my_blocks[block].read.box;
@@ -57,35 +57,35 @@ __global__ void chebyshev_kernel(level_type level, int x_id, int rhs_id, double 
       const int kStride = level.my_boxes[box].kStride;
       const int flux_jStride = (idim+1);
       const int flux_kStride = (BLOCKCOPY_TILE_J+1)*(idim+1);
-      const double * __restrict__ rhs      = level.my_boxes[box].vectors[       rhs_id] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
+      const REAL * __restrict__ rhs      = level.my_boxes[box].vectors[       rhs_id] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
 #ifdef USE_HELMHOLTZ
-      const double * __restrict__ alpha    = level.my_boxes[box].vectors[VECTOR_ALPHA ] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
+      const REAL * __restrict__ alpha    = level.my_boxes[box].vectors[VECTOR_ALPHA ] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
 #endif
-      const double * __restrict__ beta_i   = level.my_boxes[box].vectors[VECTOR_BETA_I] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
-      const double * __restrict__ beta_j   = level.my_boxes[box].vectors[VECTOR_BETA_J] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
-      const double * __restrict__ beta_k   = level.my_boxes[box].vectors[VECTOR_BETA_K] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
-      const double * __restrict__ Dinv     = level.my_boxes[box].vectors[VECTOR_DINV  ] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
+      const REAL * __restrict__ beta_i   = level.my_boxes[box].vectors[VECTOR_BETA_I] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
+      const REAL * __restrict__ beta_j   = level.my_boxes[box].vectors[VECTOR_BETA_J] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
+      const REAL * __restrict__ beta_k   = level.my_boxes[box].vectors[VECTOR_BETA_K] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
+      const REAL * __restrict__ Dinv     = level.my_boxes[box].vectors[VECTOR_DINV  ] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
 #ifdef STENCIL_FUSE_BC
-      const double * __restrict__ valid    = level.my_boxes[box].vectors[VECTOR_VALID ] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
+      const REAL * __restrict__ valid    = level.my_boxes[box].vectors[VECTOR_VALID ] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
 #endif
-            double * __restrict__ x_np1;
-      const double * __restrict__ x_n;
-      const double * __restrict__ x_nm1;
+            REAL * __restrict__ x_np1;
+      const REAL * __restrict__ x_n;
+      const REAL * __restrict__ x_nm1;
                        if((s&1)==0){x_n    = level.my_boxes[box].vectors[         x_id] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
                                     x_nm1  = level.my_boxes[box].vectors[VECTOR_TEMP  ] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
                                     x_np1  = level.my_boxes[box].vectors[VECTOR_TEMP  ] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);}
                                else{x_n    = level.my_boxes[box].vectors[VECTOR_TEMP  ] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
                                     x_nm1  = level.my_boxes[box].vectors[         x_id] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);
                                     x_np1  = level.my_boxes[box].vectors[         x_id] + ghosts*(1+jStride+kStride) + (ilo + jlo*jStride + klo*kStride);}
-      const double c1 = chebyshev_c1[s%CHEBYSHEV_DEGREE]; // limit polynomial to degree CHEBYSHEV_DEGREE.
-      const double c2 = chebyshev_c2[s%CHEBYSHEV_DEGREE]; // limit polynomial to degree CHEBYSHEV_DEGREE.
+      const REAL c1 = chebyshev_c1[s%CHEBYSHEV_DEGREE]; // limit polynomial to degree CHEBYSHEV_DEGREE.
+      const REAL c2 = chebyshev_c2[s%CHEBYSHEV_DEGREE]; // limit polynomial to degree CHEBYSHEV_DEGREE.
 
 
       int i=threadIdx.x,j=threadIdx.y,k;
       int ijk,flux_ij;
       for(k=0;k<kdim;k++){
-        double * __restrict__ flux_klo = flux_k + ((k  )&0x1)*flux_kStride;
-        double * __restrict__ flux_khi = flux_k + ((k+1)&0x1)*flux_kStride;
+        REAL * __restrict__ flux_klo = flux_k + ((k  )&0x1)*flux_kStride;
+        REAL * __restrict__ flux_khi = flux_k + ((k+1)&0x1)*flux_kStride;
 	if(bounds){
         // calculate flux_i and flux_j
               ijk = i + j*jStride + k*kStride;
@@ -154,13 +154,13 @@ __global__ void chebyshev_kernel(level_type level, int x_id, int rhs_id, double 
         // CHEBYSHEV smoother...
               ijk = i + j*jStride + k*kStride;
           flux_ij = i + j*flux_jStride;
-          double Ax = + flux_i[  flux_ij] - flux_i[  flux_ij+      1]
+          REAL Ax = + flux_i[  flux_ij] - flux_i[  flux_ij+      1]
                       + flux_j[  flux_ij] - flux_j[  flux_ij+flux_jStride]
                       + flux_klo[flux_ij] - flux_khi[flux_ij        ];
           #ifdef USE_HELMHOLTZ
           Ax += a*alpha[ijk]*x_n[ijk];
           #endif
-          double lambda =     Dinv_ijk();
+          REAL lambda =     Dinv_ijk();
 	  x_np1[ijk] = x_n[ijk] + c1*(x_n[ijk]-x_nm1[ijk]) + c2*lambda*(rhs[ijk]-Ax);
 	}
 	__syncthreads();
@@ -169,13 +169,13 @@ __global__ void chebyshev_kernel(level_type level, int x_id, int rhs_id, double 
 
 //------------------------------------------------------------------------------------------------------------------------------
 #define STENCIL_KERNEL(log_dim_i, block_i, block_j, block_k) \
-  chebyshev_kernel<log_dim_i, block_i, block_j, block_k><<<num_blocks, dim3(block_i,block_j), (block_i+1)*(BLOCKCOPY_TILE_J+1)*(4)*sizeof(double)>>>(level, x_id, rhs_id, a, b, s, chebyshev_c1, chebyshev_c2);
+  chebyshev_kernel<log_dim_i, block_i, block_j, block_k><<<num_blocks, dim3(block_i,block_j), (block_i+1)*(BLOCKCOPY_TILE_J+1)*(4)*sizeof(REAL)>>>(level, x_id, rhs_id, a, b, s, chebyshev_c1, chebyshev_c2);
 
 extern "C"
-void cuda_smooth(level_type level, int x_id, int rhs_id, double a, double b, int s, double *c, double *d)
+void cuda_smooth(level_type level, int x_id, int rhs_id, REAL a, REAL b, int s, REAL *c, REAL *d)
 {
   int num_blocks = level.num_my_blocks; if(num_blocks<=0) return;
-  int log_dim_i = (int)log2((double)level.dim.i);
+  int log_dim_i = (int)log2((REAL)level.dim.i);
   int block_dim_i = min(level.box_dim, BLOCKCOPY_TILE_I);
   int block_dim_k = min(level.box_dim, BLOCKCOPY_TILE_K);
